Issue 543: Implement Resolve{} with E-Order
This keeps strict E-Ordering and patches in the work @zenhack did to implement
`handleResolve()`. @zenhack was in the middle of working on these changes when he
was taken from us too soon. :(

After looking carefully at the commits, this keeps E-Order for calls sent over the
wire, and puts in place the unit test @zenhack wrote, `TestPromiseOrdering()`. The
original commit this was pulled from is 7e1beddbbe1ef98b922a9f81c21fe8ee97696553 on 2023-Jun-29.
In that commit @zenhack had not quite gotten the test to pass. That's a normal
intermediate stage during feature development, and it's a huge challenge to try
to measure up to @zenhack's work.

This series of commits has the two goals of -

1. Keeping strict E-Ordering because @zenhack's solution is now permanently
   unreachable.

2. Adding some comments and renaming some things as we catch up to where
   things were.

Looking at the diff, this rolls back the `resolver` field in `type impent struct {}`
found in 7e1beddbbe1ef98b922a9f81c21fe8ee97696553, in favor of a direct implementation
called `importClient.resolveNow()`.

This also renames `lockedConn.isLocalClient()` to
`lockedConn.isClientThatSendsDisembargo()` for clarity.

This commit does not include the implementation of ClientSnapshot, which was designed
to handle the Tribble 4-way race condition. The ClientSnapshot concept is still considered
good but is deferred to a later commit.